Evotree: Evolution Traversal Enhancements (#3936)
Like move validation, evolutions are the earliest thing we wish to traverse when determining what encounters may have originated the current Pokémon. To determine the permitted species-form-levels a Pokémon could originate with, we must devolve a Pokémon by traveling down-generation to origin. Once we have an encounter, we can then evolve it to the current species, traversing upwards from origin to the current format.

+++ PKHeX.Core/PersonalInfo/Interfaces/IPersonalTable.cs
@@ -8,7 +8,7 @@ public interface IPersonalTable
     /// <summary>
     /// Max Species ID (National Dex) that is stored in the table.
     /// </summary>
-    int MaxSpeciesID { get; }
+    ushort MaxSpeciesID { get; }
 
     /// <summary>
     /// Gets an index from the inner array.
