+++ PKHeX.Core/PersonalInfo/Table/PersonalTable8LA.cs
@@ -9,8 +9,8 @@ public sealed class PersonalTable8LA : IPersonalTable, IPersonalTable<PersonalIn
 {
     private readonly PersonalInfo8LA[] Table;
     private const int SIZE = PersonalInfo8LA.SIZE;
-    private const int MaxSpecies = Legal.MaxSpeciesID_8a;
-    public int MaxSpeciesID => MaxSpecies;
+    private const ushort MaxSpecies = Legal.MaxSpeciesID_8a;
+    public ushort MaxSpeciesID => MaxSpecies;
 
     public PersonalTable8LA(ReadOnlySpan<byte> data)
     {
